+++ misosoup/reframed/gurobi_env_solver.py
@@ -1,0 +1,20 @@
+"""Gurobi solver instance for individual environments."""
+
+from reframed.solvers.solver import Solver
+from reframed.solvers.gurobi_solver import GurobiSolver, default_parameters
+from gurobipy import Model as GurobiModel
+
+
+class GurobiEnvSolver(GurobiSolver):
+    """Gurobi interface initialized in gurobi environment."""
+
+    def __init__(self, model=None, env=None):
+        Solver.__init__(self)
+        if env:
+            self.problem = GurobiModel(env=env)
+        else:
+            self.problem = GurobiModel()
+        self.set_logging(False)
+        self.set_parameters(default_parameters)
+        if model:
+            self.build_problem(model)